[BAHIR-90] Include Apache Flink extensions to website
This commit setups links for documentation and downloads
for Apache Bahir Flink extensions. The Flink docs are
generated using the updated update-docs.sh script in #8.

It also revamps the website's project infrastructure,
configuration, and navigation / layouts for documentations
and the download page to be more friendly for newly
supported platforms in the future.

Closes #9